Separado código de abstracción con ruby en varios ficheros
